requirements.txt
bcrypt==4.2.1
blinker==1.9.0
click==8.1.7
dnspython==2.7.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pymongo==4.10.1
Werkzeug==3.1.3
python-dotenv==1.0.0